Добавил pom.xml

+++ src/test/java/LoginUserTest.java
@@ -1,3 +1,4 @@
+import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
 import org.example.steps.StepsToCreateUser;
@@ -9,9 +10,9 @@ import org.junit.Test;
 
 public class LoginUserTest {
 
-    private final String email = "5ssdf8@31.ru";
-    private final String password = "1234";
-    private final String name = "aboba";
+    private final String email = BaseClass.email();
+    private final String password = BaseClass.password();
+    private final String name = BaseClass.name();
 
     @Before
     public void setUp() {
@@ -20,7 +21,8 @@ public class LoginUserTest {
 
     @Test
     @DisplayName("Авторизация пользователя")
-    public void loginUser() {
+    @Description("")
+    public void loginUserTest() {
         StepsToCreateUser.createUserInOneStep(email, password, name);
         Response response = StepsToLoginUser.loginUser(email, password);
         StepsToLoginUser.checkStatusCodeWhenLoginUser(response);
@@ -29,7 +31,8 @@ public class LoginUserTest {
 
     @Test
     @DisplayName("Невозможно авторизоваться без логина")
-    public void cantLoginUserWithoutEmail() {
+    @Description("Проверка, что при попытке авторизоваться без логина код ответа равен 401 и значение 'success' равно 'false'")
+    public void cantLoginUserWithoutEmailTest() {
         StepsToCreateUser.createUserInOneStep(email, password, name);
         Response response = StepsToLoginUser.loginUser("", password);
         StepsToLoginUser.checkStatusCodeWhenLoginUserWithoutLoginOrPassword(response);
@@ -37,17 +40,40 @@ public class LoginUserTest {
     }
 
     @Test
+    @DisplayName("Невозможно авторизоваться с неверным логином")
+    @Description("Проверка, что при попытке авторизоваться с неверным логином код ответа равен 401 и значение 'success' равно 'false'")
+    public void cantLoginUserWithWrongEmailTest() {
+        StepsToCreateUser.createUserInOneStep(email, password, name);
+        Response response = StepsToLoginUser.loginUser("1234qwerty@mail.ru", password);
+        StepsToLoginUser.checkStatusCodeWhenLoginUserWithoutLoginOrPassword(response);
+        StepsToLoginUser.checkResponseWhenLoginUserWithoutLoginOrPassword(response);
+    }
+
+    @Test
     @DisplayName("Невозможно авторизоваться без пароля")
-    public void cantLoginUserWithoutPassword() {
+    @Description("Проверка, что при попытке авторизоваться без пароля код ответа равен 401 и значение 'success' равно 'false'")
+    public void cantLoginUserWithoutPasswordTest() {
         StepsToCreateUser.createUserInOneStep(email, password, name);
         Response response = StepsToLoginUser.loginUser(email, "");
         StepsToLoginUser.checkStatusCodeWhenLoginUserWithoutLoginOrPassword(response);
         StepsToLoginUser.checkResponseWhenLoginUserWithoutLoginOrPassword(response);
     }
 
+    @Test
+    @DisplayName("Невозможно авторизоваться с неверным паролем")
+    @Description("Проверка, что при попытке авторизоваться с неверным паролем код ответа равен 401 и значение 'success' равно 'false'")
+    public void cantLoginUserWithWrongPasswordTest() {
+        StepsToCreateUser.createUserInOneStep(email, password, name);
+        Response response = StepsToLoginUser.loginUser(email, password + "1");
+        StepsToLoginUser.checkStatusCodeWhenLoginUserWithoutLoginOrPassword(response);
+        StepsToLoginUser.checkResponseWhenLoginUserWithoutLoginOrPassword(response);
+    }
+
     @After
     public void tearDown() {
-        StepsToDeleteUser.deleteUser(StepsToLoginUser.loginUserInOneStep(email, password));
+        if (StepsToLoginUser.loginUser(email, password).statusCode() != 401) {
+            StepsToDeleteUser.deleteUser(StepsToLoginUser.loginUserInOneStep(email, password));
+        }
     }
 }
